chore: update teaser media docs
