+++ packages/components/src/components/teaser-media/teaser-media.stories.ts
@@ -1,259 +1,146 @@
 import '../../solid-components';
 import { html } from 'lit';
-import { storybookDefaults, storybookHelpers, storybookTemplate } from '../../../scripts/storybook/helper';
+import { storybookDefaults } from '../../../scripts/storybook/helper';
 
 const { argTypes, parameters } = storybookDefaults('sd-teaser-media');
-const { overrideArgs } = storybookHelpers('sd-teaser-media');
-const { generateTemplate } = storybookTemplate('sd-teaser-media');
 
+/**
+ * Flexible containers that group related items and link to further content.
+ *
+ * **Related templates**:
+ * - [Teaser Media](?path=/docs/templates-teaser-media--docs)
+ */
 export default {
+  tags: ['!dev'],
   title: 'Components/sd-teaser-media',
   component: 'sd-teaser-media',
-  args: overrideArgs([
-    {
-      type: 'slot',
-      name: 'default',
-      value: `<div class="slot slot--border slot--text h-12">Main slot</div>`
-    },
-    {
-      type: 'slot',
-      name: 'media',
-      value: `<div slot="media" class="slot slot--border slot--text h-[384px]">Media slot</div>`
-    },
-    {
-      type: 'slot',
-      name: 'meta',
-      value: `<div slot="meta" class="slot slot--border slot--text h-12">Meta slot</div>`
-    }
-  ]),
   argTypes,
   parameters
 };
 
-/**
- * This shows sd-teaser-media in its default state.
- */
-
 export const Default = {
-  render: (args: any) => {
-    return generateTemplate({ args });
-  }
+  name: 'Default',
+  render: () =>
+    html`<sd-teaser-media variant="white">
+      <div class="slot slot--border slot--text h-12">Main slot</div>
+      <div slot="media" class="slot slot--border slot--text h-[384px]">Media slot</div>
+      <div slot="meta" class="slot slot--border slot--text h-12">Meta slot</div>
+    </sd-teaser-media>`
 };
 
 /**
- * Teaser-Media in all possible `variant`.
+ * Use the `variant` attribute to set the color variant:
+ * - `white (default)`
+ * - `gradient-dark`
+ * - `gradient-white`
+ * - `primary`
+ * - `primary-100`
+ * - `neutral-100`
  */
 
 export const Variant = {
   name: 'Variant',
-  parameters: { controls: { exclude: ['variant'] } },
-  render: (args: any) => {
-    return generateTemplate({
-      axis: {
-        y: { type: 'attribute', name: 'variant' }
-      },
-      args,
-      constants: [
-        {
-          type: 'template',
-          name: 'style',
-          value: '<div style="margin-bottom: 40px; width: 600px">%TEMPLATE%</div>'
-        },
-        {
-          type: 'slot',
-          name: 'media',
-          value:
-            '<img slot="media" class="aspect-video object-cover" src="./placeholders/images/generic.jpg" alt="A skyline of a city by night"/>'
-        }
-      ]
-    });
-  }
-};
+  render: () => html`
+    <div style="margin-bottom: 40px; width: 600px">
+      <sd-teaser-media variant="white">
+        <img slot="media" src="./placeholders/images/collaboration.jpg" alt="Test" />
+        <div slot="meta">
+          <span class="meta-info-item text-sm">01.12.2013</span>
+          <span class="meta-info-item text-sm">| Author name</span>
+        </div>
+        <h3 slot="headline">Headline Media Teaser (white – default)</h3>
+        <div class="flex flex-col gap-4">
+          <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</p>
+          <div class="flex-none">
+            <sd-button href="#" target="_blank" variant="primary">Label</sd-button>
+          </div>
+        </div>
+      </sd-teaser-media>
+    </div>
 
-/**
- * Use the 'default', 'media', 'meta', 'expandable' and 'headline' slots to add content to the teaser. Please make sure to use semantically correct headline tags for the `headline` slot to provide accessible content.
- */
-export const Slots = {
-  parameters: {
-    controls: { exclude: ['default', 'media', 'meta', 'headline', 'expandable'] }
-  },
-  render: (args: any) => {
-    return html`
-      ${['default', 'media', 'meta', 'headline', 'expandable'].map(slot => {
-        let value = `<div slot='${slot}' class="slot slot--border slot--background h-16"></div>`;
+    <div style="margin-bottom: 40px; width: 600px">
+      <sd-teaser-media variant="gradient-dark">
+        <img slot="media" src="./placeholders/images/collaboration.jpg" alt="Test" />
+        <div slot="meta">
+          <span class="meta-info-item text-sm">01.12.2013</span>
+          <span class="meta-info-item text-sm">| Author name</span>
+        </div>
+        <h3 slot="headline">Headline Media Teaser (gradient-dark)</h3>
+        <div class="flex flex-col gap-4">
+          <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</p>
+          <div class="flex-none">
+            <sd-button href="#" target="_blank" variant="primary" inverted>Label</sd-button>
+          </div>
+        </div>
+      </sd-teaser-media>
+    </div>
 
-        if (slot === 'default') {
-          value = `<div class="slot slot--border slot--background h-16"></div>`;
-        } else if (slot === 'media') {
-          value = `<div slot='${slot}' class="slot slot--border slot--background h-[577px]"></div>`;
-        } else if (slot === 'expandable') {
-          value = `<div slot='${slot}' class="slot--border slot--background h-16 box-border flex flex-col items-center justify-center rounded-md"></div>`;
-        }
+    <div style="margin-bottom: 40px; width: 600px">
+      <sd-teaser-media variant="gradient-white">
+        <img slot="media" src="./placeholders/images/collaboration.jpg" alt="Test" />
+        <div slot="meta">
+          <span class="meta-info-item text-sm">01.12.2013</span>
+          <span class="meta-info-item text-sm">| Author name</span>
+        </div>
+        <h3 slot="headline">Headline Media Teaser (gradient-white)</h3>
+        <div class="flex flex-col gap-4">
+          <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</p>
+          <div class="flex-none">
+            <sd-button href="#" target="_blank" variant="primary">Label</sd-button>
+          </div>
+        </div>
+      </sd-teaser-media>
+    </div>
 
-        return generateTemplate({
-          axis: {
-            x: {
-              type: 'slot',
-              name: slot,
-              title: 'slot=..',
-              values: [
-                {
-                  value: value,
-                  title: slot
-                }
-              ]
-            }
-          },
-          args,
-          constants: [
-            {
-              type: 'slot',
-              name: 'media',
-              value: `<img slot='media' src='./placeholders/images/collaboration.jpg' alt='Test'/>`
-            },
-            {
-              type: 'slot',
-              name: 'meta',
-              value: `<slot slot='meta'>Teaser's Meta information</slot>`
-            },
-            {
-              type: 'slot',
-              name: 'default',
-              value: `<slot>Teaser's Main content</slot>`
-            },
-            {
-              type: 'slot',
-              name: 'headline',
-              value: `<slot slot='headline'>Teaser's Headline</slot>`
-            },
-            {
-              type: 'slot',
-              name: 'expandable',
-              value: `<slot slot='expandable'>Teaser's Expandable content</slot>`
-            },
-            {
-              type: 'template',
-              name: 'style',
-              value: '<div style="margin-bottom: 40px; width: 782px;">%TEMPLATE%</div>'
-            }
-          ]
-        });
-      })}
-    `;
-  }
-};
+    <div style="margin-bottom: 40px; width: 600px">
+      <sd-teaser-media variant="primary">
+        <img slot="media" src="./placeholders/images/collaboration.jpg" alt="Test" />
+        <div slot="meta">
+          <span class="meta-info-item text-sm">01.12.2013</span>
+          <span class="meta-info-item text-sm">| Author name</span>
+        </div>
+        <h3 slot="headline">Headline Media Teaser (primary)</h3>
+        <div class="flex flex-col gap-4">
+          <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</p>
+          <div class="flex-none">
+            <sd-button href="#" target="_blank" variant="primary" inverted>Label</sd-button>
+          </div>
+        </div>
+      </sd-teaser-media>
+    </div>
 
-export const Parts = {
-  parameters: {
-    controls: { exclude: ['base', 'media', 'content', 'meta', 'headline', 'main', 'expandable', 'variant'] }
-  },
-  render: (args: any) => {
-    return generateTemplate({
-      axis: {
-        y: {
-          type: 'template',
-          name: 'sd-teaser-media::part(...){outline: solid 2px red}',
-          values: ['base', 'media', 'content', 'meta', 'headline', 'main', 'expandable'].map(part => {
-            return {
-              title: part,
-              value: `<style>#part-${part} sd-teaser-media::part(${part}){outline: solid 2px red}</style><div id="part-${part}">%TEMPLATE%</div>`
-            };
-          })
-        }
-      },
-      args,
-      constants: [
-        {
-          type: 'template',
-          name: 'style',
-          value: '<div style="margin-bottom: 40px; width: 500px">%TEMPLATE%</div>'
-        },
-        {
-          type: 'attribute',
-          name: 'variant',
-          value: 'white'
-        },
-        {
-          type: 'slot',
-          name: 'expandable',
-          value: `<div slot="expandable" class="slot--border slot--text box-border flex flex-col items-center justify-center rounded-md"><p>Expandable slot</p></div>`
-        }
-      ]
-    });
-  }
-};
+    <div style="margin-bottom: 40px; width: 600px">
+      <sd-teaser-media variant="primary-100">
+        <img slot="media" src="./placeholders/images/collaboration.jpg" alt="Test" />
+        <div slot="meta">
+          <span class="meta-info-item text-sm">01.12.2013</span>
+          <span class="meta-info-item text-sm">| Author name</span>
+        </div>
+        <h3 slot="headline">Headline Media Teaser (primary-100)</h3>
+        <div class="flex flex-col gap-4">
+          <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</p>
+          <div class="flex-none">
+            <sd-button href="#" target="_blank" variant="primary">Label</sd-button>
+          </div>
+        </div>
+      </sd-teaser-media>
+    </div>
 
-export const Samples = {
-  name: 'Samples: Teaser-Media',
-  parameters: {
-    controls: {
-      disable: true
-    },
-    backgrounds: {
-      default: 'white'
-    }
-  },
-  render: () => {
-    return html`
-      <style>
-        #teaserWithContentPlaceholder::part(media) {
-          flex-grow: 1;
-        }
-      </style>
-      <div class="flex flex-wrap gap-8">
-        <sd-teaser-media variant="gradient-dark" class="min-w-[435px] max-w-4xl">
-          <div slot="media" class="relative">
-            <img
-              class="aspect-video object-cover"
-              src="./placeholders/images/architecture.jpg"
-              alt="A skyline of a city"
-            />
+    <div style="margin-bottom: 40px; width: 600px">
+      <sd-teaser-media variant="neutral-100">
+        <img slot="media" src="./placeholders/images/collaboration.jpg" alt="Test" />
+        <div slot="meta">
+          <span class="meta-info-item text-sm">01.12.2013</span>
+          <span class="meta-info-item text-sm">| Author name</span>
+        </div>
+        <h3 slot="headline">Headline Media Teaser (neutral-100)</h3>
+        <div class="flex flex-col gap-4">
+          <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</p>
+          <div class="flex-none">
+            <sd-button href="#" target="_blank" variant="primary">Label</sd-button>
           </div>
-          <div slot="meta" class="meta-info">
-            <span class="meta-info-item">01.12.2013</span>
-            <span class="meta-info-item">| Author name</span>
-          </div>
-          <h3 slot="headline">Not expandable teaser-media</h3>
-          <div class="flex flex-col gap-4">
-            <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</p>
-            <div class="flex-none">
-              <sd-button href="#" target="_blank" variant="primary" inverted>Link</sd-button>
-            </div>
-          </div>
-        </sd-teaser-media>
-        <sd-teaser-media variant="gradient-white" class="min-w-[435px] max-w-4xl">
-          <div slot="media" class="relative">
-            <img
-              class="aspect-video object-cover"
-              src="./placeholders/images/architecture.jpg"
-              alt="A skyline of a city"
-            />
-          </div>
-          <div slot="meta" class="meta-info">
-            <span class="meta-info-item">01.12.2013</span>
-            <span class="meta-info-item">| Author name</span>
-          </div>
-          <h3 slot="headline">Expandable teaser-media</h3>
-          <div slot="expandable">
-            <p>
-              Expandable: Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut
-              labore et dolore magna aliqua.
-            </p>
-          </div>
-          <div class="flex flex-col gap-4">
-            <p>
-              Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et
-              dolore magna aliqua.
-            </p>
-            <div class="flex-none">
-              <sd-button href="#" target="_blank" variant="primary">Link</sd-button>
-            </div>
-            <div class="flex-none text-sm">
-              <p>@Copyright Lorem ipsum</p>
-            </div>
-          </div>
-        </sd-teaser-media>
-      </div>
-    `;
-  }
+        </div>
+      </sd-teaser-media>
+    </div>
+  `
 };
